add webpack merge
