+++ public/index.html
@@ -1,43 +1,100 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./styles.css">
-    <title>One click meals</title>
-
+    <title>Buscador de recetas</title>
+    <link rel="stylesheet" href="../src/css/styles.css">
 </head>
-
 <body>
-    <section class="fixed-header">
-        <div class="container">
-            <h1>One click meals</h1>
-            <div class="flex">
-                <form class="flex" id="submit">
-                    <input type="text" id="search" placeholder="Search for meals or keywords" />
-                    <button class="search-btn" type="submit">
-                        <i class="fas fa-search"></i>
-                    </button>
-                </form>
-            </div>
-            <div>
-                <button class="random-btn" id="random">
-                    <i class="fas fa-random"></i>
-                </button>
-            </div>
+    <div class="dish-selected-content no-display">
+        <div class="dish-selected">
+            <button id="btnCloseDishSelected" class="button">
+                <svg xmlns="http://www.w3.org/2000/svg" class="icon-close" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#000000" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                    <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                    <line x1="12" y1="5" x2="12" y2="19" />
+                    <line x1="5" y1="12" x2="19" y2="12" />
+                  </svg>
+            </button>
         </div>
-    </section>
-    <div class="results">
-        <div id="result-heading"></div>
-        <div id="meals" class="meals"></div>
-        <div id="single-meal"></div>
+        <div class="dish-selected-background"></div>
     </div>
-
-
-    <script src="https://kit.fontawesome.com/31ac679c64.js" crossorigin="anonymous"></script>
-    <script src="/src/js/delete.js"></script>
+    <header>
+        <nav class="nav">
+            <ul class="navigation">
+                <li class="navigation__list">
+                    <svg xmlns="http://www.w3.org/2000/svg" class="icon-menu" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#07090D" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                        <line x1="4" y1="6" x2="20" y2="6" />
+                        <line x1="4" y1="12" x2="20" y2="12" />
+                        <line x1="4" y1="18" x2="20" y2="18" />
+                    </svg>
+                </li>
+                <li class="navigation__list">
+                    <div class="user-location">
+                        <svg xmlns="http://www.w3.org/2000/svg" class="icon-location" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#ffffff" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                            <circle cx="12" cy="11" r="3" />
+                            <path d="M17.657 16.657l-4.243 4.243a2 2 0 0 1 -2.827 0l-4.244 -4.243a8 8 0 1 1 11.314 0z" />
+                          </svg>
+                        <span>México</span>
+                    </div>
+                </li>
+                <li class="navigation__list shopping-cart">
+                    <span class="shopping-cart__num-items">3</span>
+                    <svg xmlns="http://www.w3.org/2000/svg" class="icon-shopping-cart" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#07090D" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                        <circle cx="6" cy="19" r="2" />
+                        <circle cx="17" cy="19" r="2" />
+                        <path d="M17 17h-11v-14h-2" />
+                        <path d="M6 5l14 1l-1 7h-13" />
+                    </svg>
+                </li>
+            </ul>
+        </nav>
+    </header>
+    <main>
+        <section class="search-content">
+            <h1>¡Recetas de todo el mundo!</h1>
+            <form action="" id="search">
+                <div>
+                    <svg xmlns="http://www.w3.org/2000/svg" class="icon-search" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#363637" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                        <circle cx="10" cy="10" r="7" />
+                        <line x1="21" y1="21" x2="15" y2="15" />
+                      </svg>
+                    <input type="text" name="" id="txtRecipe" placeholder="Busca una receta" required>
+                    <button id="btnSearchRandom">
+                        <svg xmlns="http://www.w3.org/2000/svg" class="icon-random" width="44" height="44" viewBox="0 0 24 24" stroke-width="1.5" stroke="#000000" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                            <polyline points="15 4 19 4 19 8" />
+                            <line x1="14.75" y1="9.25" x2="19" y2="4" />
+                            <line x1="5" y1="19" x2="9" y2="15" />
+                            <polyline points="15 19 19 19 19 15" />
+                            <line x1="5" y1="5" x2="19" y2="19" />
+                          </svg>
+                    </button>
+                </div>
+                <button id="btnSearchRecipe">Buscar</button>
+            </form>
+            <div id="spinner">
+                <div class="sk-chase">
+                    <div class="sk-chase-dot"></div>
+                    <div class="sk-chase-dot"></div>
+                    <div class="sk-chase-dot"></div>
+                    <div class="sk-chase-dot"></div>
+                    <div class="sk-chase-dot"></div>
+                    <div class="sk-chase-dot"></div>
+                  </div>
+            </div>
+        </section>
+        <section class="dishes-content">
+            <div class="dishes">
+            </div>
+        </section>
+    </main>
+      
+    <script src="../src/js/index.js"></script>
 </body>
-
 </html>
